==> 01_Build_Podcast_List.ipynb <==
%matplotlib inline

import pandas as pd
import seaborn as sns
import numpy as np
import guess_language

# load all .tsv files
# source: https://github.com/ageitgey/all-podcasts-dataset
colNames = ['slug', 'name', 'image_url', 'feed_url', 'website_url', 'itunes_owner_name', 'itunes_owner_email', 'managing_editor_name', 'managing_editor_email', 'explicit', 'description', 'itunes_summary'] 
podcastDf = pd.DataFrame(columns=colNames)
files = !ls data/*.tsv

for file in files:
    podcastDf = podcastDf.append(pd.read_table(file, header=None, names=colNames), ignore_index=True)

# remove podcasts that don't have itunes summary or itunes explicit rating (probably not in itunes)
# remove podcasts that don't have feed url
podcastDf = podcastDf.dropna(subset=['itunes_summary', 'explicit', 'feed_url'])

# guess language based on itunes_summary
lang_guess = []
for x in podcastDf['itunes_summary']:
    lang_guess.append(guess_language.guessLanguageName(x))

podcastDf['language'] = pd.Series(lang_guess, index = podcastDf.index)

# keep only English podcasts
podcastDf = podcastDf.query("language == 'English'")

# save df to pickle
podcastDf.to_pickle('pkl/english_podcasts.pkl')
