==> 02_v3_iTunes_Search_API.ipynb <==
import pandas as pd
import urllib2
import requests
import json
import time
import tqdm
from BeautifulSoup import BeautifulSoup
from common_func import convert_unicode, check_url

# load df of english language podcasts
podcastDf = pd.read_pickle('pkl/english_podcasts.pkl')

# construct base search url, later we'll add the podcast name
itunesUrl = 'https://itunes.apple.com/search?'
parameters = {}
parameters['entity'] = 'podcast'
parameters['attribute'] = 'titleTerm'

requestString = ''
for param in parameters:
    paramString = param + '=' + parameters[param]
    if len(requestString) == 0:
        requestString = paramString
    else:
        requestString = requestString + '&' + paramString

itunesUrl = itunesUrl + requestString + '&term='
print itunesUrl

# make subset of podcastDf for testing
testDf = podcastDf.sample(n=20)

# do a test request so we get the column names
testUrl = itunesUrl + urllib2.quote('political gabfest')
r = requests.get(testUrl)
text = r.text
data = json.loads(text)
thisResult = pd.io.json.json_normalize(data['results'])
thisResult['id'] = 12345
colNames = list(thisResult.columns.values)

searchResults = pd.DataFrame(columns=colNames)
searchResults

# set parameters for checking API search url
num_tries = 4
wait_secs = 60 # interval between tries

for ind, row in tqdm.tqdm(podcastDf.iterrows(), total=podcastDf.shape[0]):
    
    # check if url works
    searchUrl = itunesUrl + urllib2.quote(row['name'])
    url_error = check_url(searchUrl, num_tries, wait_secs)
    if url_error:
        print url_error
        break
    
    # get search results
    r = requests.get(searchUrl)
    text = r.text
    data = json.loads(text)
    if data['resultCount'] != 1:
        thisResult = pd.DataFrame({'id' : [ind]})
        searchResults = pd.concat([searchResults, thisResult], axis=0)
    else:
        thisResult = pd.io.json.json_normalize(data['results'])
        thisResult['id'] = ind
        searchResults = pd.concat([searchResults, thisResult], axis=0)

# pickle the current results
searchResults.to_pickle('pkl/itunes_podcasts_v3_pt1.pkl')

# remove already-retrieved values from podcastDf
doneInds = searchResults['id']
allInds = podcastDf.index
notDoneInds = allInds.sym_diff(doneInds)

subDf = podcastDf.loc[notDoneInds]

# set parameters for checking API search url
num_tries = 4
wait_secs = 60 # interval between tries

for ind, row in tqdm.tqdm(subDf.iterrows(), total=subDf.shape[0]):
    
    # check if url works
    searchUrl = itunesUrl + urllib2.quote(row['name'])
    url_error = check_url(searchUrl, num_tries, wait_secs)
    if url_error:
        print url_error
        break
    
    # get search results
    r = requests.get(searchUrl)
    text = r.text
    data = json.loads(text)
    if data['resultCount'] != 1:
        thisResult = pd.DataFrame({'id' : [ind]})
        searchResults = pd.concat([searchResults, thisResult], axis=0)
    else:
        thisResult = pd.io.json.json_normalize(data['results'])
        thisResult['id'] = ind
        searchResults = pd.concat([searchResults, thisResult], axis=0)

searchResults.shape

# pickle the current results
searchResults.to_pickle('pkl/itunes_podcasts_v3_pt2.pkl')

# remove already-retrieved values from podcastDf
doneInds = searchResults['id']
allInds = podcastDf.index
notDoneInds = allInds.sym_diff(doneInds)

subDf = podcastDf.loc[notDoneInds]

# set parameters for checking API search url
num_tries = 4
wait_secs = 60 # interval between tries

for ind, row in tqdm.tqdm(subDf.iterrows(), total=subDf.shape[0]):
    
    # check if url works
    searchUrl = itunesUrl + urllib2.quote(row['name'])
    url_error = check_url(searchUrl, num_tries, wait_secs)
    if url_error:
        print url_error
        break
    
    # get search results
    r = requests.get(searchUrl)
    text = r.text
    data = json.loads(text)
    if data['resultCount'] != 1:
        thisResult = pd.DataFrame({'id' : [ind]})
        searchResults = pd.concat([searchResults, thisResult], axis=0)
    else:
        thisResult = pd.io.json.json_normalize(data['results'])
        thisResult['id'] = ind
        searchResults = pd.concat([searchResults, thisResult], axis=0)

# pickle the current results
searchResults.to_pickle('pkl/itunes_podcasts_v3_pt3.pkl')

# remove already-retrieved values from podcastDf
doneInds = searchResults['id']
allInds = podcastDf.index
notDoneInds = allInds.sym_diff(doneInds)

subDf = podcastDf.loc[notDoneInds]

# set parameters for checking API search url
num_tries = 4
wait_secs = 60 # interval between tries

for ind, row in tqdm.tqdm(subDf.iterrows(), total=subDf.shape[0]):
    
    # check if url works
    searchUrl = itunesUrl + urllib2.quote(row['name'])
    url_error = check_url(searchUrl, num_tries, wait_secs)
    if url_error:
        print url_error
        break
    
    # get search results
    r = requests.get(searchUrl)
    text = r.text
    data = json.loads(text)
    if data['resultCount'] != 1:
        thisResult = pd.DataFrame({'id' : [ind]})
        searchResults = pd.concat([searchResults, thisResult], axis=0)
    else:
        thisResult = pd.io.json.json_normalize(data['results'])
        thisResult['id'] = ind
        searchResults = pd.concat([searchResults, thisResult], axis=0)

searchResults.shape

# pickle the current results
searchResults.to_pickle('pkl/itunes_podcasts_v3_COMPLETE.pkl')
