==> 04_Clean_Data.ipynb <==
import pandas as pd
import re
import guess_language

# load data
scrapeDf = pd.read_pickle('pkl/scraped_podcasts_COMPLETE.pkl')
apiDf = pd.read_pickle('pkl/itunes_podcasts_v3_COMPLETE.pkl')

# clean column names because SQL doesn't like camelCase
def convert(name):
    """Converts camelCase to camel_case."""
    s1 = re.sub('(.)([A-Z][a-z]+)', r'\1_\2', name)
    return re.sub('([a-z0-9])([A-Z])', r'\1_\2', s1).lower()

scrapeDf.columns = [convert(x) for x in scrapeDf.columns]
apiDf.columns = [convert(x) for x in apiDf.columns]

podcastDf = pd.merge(apiDf, scrapeDf, how = 'inner', on = 'collection_id')

podcastDf.shape

# convert to int
podcastDf['collection_id'] = [int(x) for x in podcastDf['collection_id']]

# guess language of summary
podcastDf['language'] = [guess_language.guessLanguageName(x) for x in podcastDf['podcast_summary']]
podcastDf = podcastDf[podcastDf['language'] == 'English']
podcastDf.shape

podcastDf.columns

episodeDf = podcastDf[['collection_id','episode_descriptions', 'episode_names']]

# convert to list
episodeDf['episode_descriptions'] = [x.split("\',") for x in episodeDf['episode_descriptions']]
episodeDf['episode_names'] = [x.split("\',") for x in episodeDf['episode_names']]

test_description = episodeDf.iloc[0:10]['episode_descriptions']

def clean_text(text):
    # remove mixed alphanumeric
    text = re.sub(r"""(?x) # verbose regex
                            \b    # Start of word
                            (?=   # Look ahead to ensure that this word contains...
                             \w*  # (after any number of alphanumeric characters)
                             \d   # ...at least one digit.
                            )     # End of lookahead
                            \w+   # Match the alphanumeric word
                            \s*   # Match any following whitespace""", 
                             "", text)
    
    # remove urls
    text = re.sub(r'\s([\S]*.com[\S]*)\b', '', text)
    text = re.sub(r'\s([\S]*.org[\S]*)\b', '', text)

    
    
    # remove non-alphanumeric, non-space
    text = re.sub(r'([^\s\w]|_)+', '', text)
    
    
    
    return text

clean_episode_description = []
for pod in episodeDf['episode_descriptions']:
    clean_list = []
    for ep in pod:
        clean_list.append(clean_text(ep))
    clean_episode_description.append(clean_list)

#episodeDf['episode_descriptions'] = [clean_text(ep) for pod in episodeDf['episode_descriptions'] for ep in pod]
    

episodeDf['clean_episode_description'] = clean_episode_description

clean_episode_name = []
for pod in episodeDf['episode_names']:
    clean_list = []
    for ep in pod:
        clean_list.append(clean_text(ep))
    clean_episode_name.append(clean_list)

episodeDf['clean_episode_name'] = clean_episode_name

episodeDf.head()

del episodeDf['episode_descriptions']
del episodeDf['episode_names']

podcastDf.columns

del podcastDf['episode_descriptions']
del podcastDf['episode_names']

podcastDf = pd.merge(podcastDf, episodeDf, how = 'inner', on='collection_id')

# clean podcast summary
podcastDf['podcast_summary'] = [clean_text(x) for x in podcastDf['podcast_summary']]

# save clean data
podcastDf.to_pickle('pkl/clean_podcast_data.pkl')
