==> 04_Make_SQL_Database.ipynb <==
from sqlalchemy import create_engine
from sqlalchemy_utils import database_exists, create_database
import psycopg2
import pandas as pd
import re

dbname = 'podcast'
username = 'lindsay'

# create connection to database
engine = create_engine('postgres://%s@localhost/%s'%(username, dbname))
print engine.url

# create database
if not database_exists(engine.url):
    create_database(engine.url)
print(database_exists(engine.url))

scrapeDf = pd.read_pickle('pkl/scraped_podcasts_COMPLETE.pkl')
apiDf = pd.read_pickle('pkl/itunes_podcasts_v3_COMPLETE.pkl')

# convert camelCase to camel_case
def convert(name):
    """Converts camelCase to camel_case."""
    s1 = re.sub('(.)([A-Z][a-z]+)', r'\1_\2', name)
    return re.sub('([a-z0-9])([A-Z])', r'\1_\2', s1).lower()

scrapeDf.columns = [convert(x) for x in scrapeDf.columns]
apiDf.columns = [convert(x) for x in apiDf.columns]

# join dfs
podcastDf = pd.merge(scrapeDf, apiDf, how='inner', on='collection_id')

# insert data into database
podcastDf.to_sql('podcasts', engine, if_exists='replace')

# connect for querying
con = None
con = psycopg2.connect(database = dbname, user = username)

# query
sql_query = """
SELECT * FROM podcasts WHERE collection_id=515836681;
"""

query_result = pd.read_sql_query(sql_query, con)
query_result
