==> 07_Create_Corpus_LDA.ipynb <==
from gensim import corpora, models, similarities
from collections import defaultdict
import re
import pandas as pd
import logging

logging.basicConfig(format='%(asctime)s : %(levelname)s : %(message)s', level=logging.INFO)

documents = pd.read_pickle('gensim/preprocessed_text_v3.pkl')

dictionary = corpora.Dictionary.load('gensim/dictionary.dict')
corpus_tfidf = corpora.MmCorpus('gensim/corpus_tfidf.mm')

# lda model on tfidf
%time lda = models.LdaModel(corpus_tfidf, num_topics = 100, id2word=dictionary)
lda.save('gensim/model.lda')

# calculate similarities
index = similarities.MatrixSimilarity(lda[corpus_tfidf])
index.save('gensim/tfidf_lda_similarities.index')
