==> 08_Calculate_Similarity.ipynb <==
from gensim import corpora, models, similarities
import pandas as pd
import tqdm
import logging

logging.basicConfig(format='%(asctime)s : %(levelname)s : %(message)s', level=logging.INFO)

# load gensim output
dictionary = corpora.Dictionary.load('gensim/dictionary.dict')
corpus = corpora.MmCorpus('gensim/corpus_tfidf.mm')
lsi = models.LsiModel.load('gensim/model.lsi')
index = similarities.MatrixSimilarity.load('gensim/tfidf_lsi_similarities.index')

# load podcast data
podcast_data = pd.read_pickle('gensim/preprocessed_text_v2.pkl')
podcast_data.shape

podcast_data = podcast_data.reset_index()
del(podcast_data['index'])
podcast_data.head()

podcast_data.to_pickle('gensim/preprocessed_text_v3.pkl')

podcast_ids = podcast_data.copy(deep=True)
del(podcast_ids['language'])
del(podcast_ids['text'])


podcast_ids.to_pickle('gensim/podcast_id_to_gensim_id.pkl')

podcast_ids = pd.read_pickle('gensim/podcast_id_to_gensim_id.pkl')

podcast_ids.head()

podcast_id = 17233
row_id = podcast_data[podcast_data['id'] == podcast_id].index
row_id = row_id[0]

row_id

search_vec = lsi[corpus[row_id]]
sims = index[search_vec]
#sims = sorted(enumerate(sims), key=lambda item: -item[1])

sims

#match_ids = [x[0] for x in sims]

sim_df = pd.DataFrame({"similarity" : sims})
sim_df.head()

podcast_ids.columns = ['podcast_id']

sim_df = pd.concat([sim_df, podcast_ids], axis = 1)

sim_df = sim_df.sort_values('similarity', ascending=False)

sim_df[sim_df['podcast_id'] != podcast_id]

sim_df_no_self = sim_df[sim_df['podcast_id'] != podcast_id]

# take top ids
ids = list(sim_df_no_self['podcast_id'].values)
ids = [str(int(x)) for x in ids]
ids = ids[:100]

import psycopg2
user = 'lindsay'          
host = 'localhost'
dbname = 'podcast'
con = None
con = psycopg2.connect(database = dbname, user = user)
cursor = con.cursor()

def pg_int_array(the_list):
  return '(' + ','.join(the_list) + ')'

query = """
SELECT name, view_url, artwork_url100, id
FROM podcast
WHERE id IN %s;
"""
query = query.replace('\n', ' ')

cursor.execute(query % (pg_int_array(ids)))
query_results = cursor.fetchall()

columnNames = ['name', 'view_url', 'artwork_url100', 'id']
podcast_results = pd.DataFrame(columns=columnNames)

podcast_results['name'] = [x[0] for x in query_results]
podcast_results['view_url'] = [x[1] for x in query_results]
podcast_results['artwork_url100'] = [x[2] for x in query_results]
podcast_results['id'] = [x[3] for x in query_results]
podcast_results.head()

merge_results = pd.merge(sim_df_no_self, podcast_results, how = 'inner', left_on='podcast_id', right_on='id')

merge_results.head()

similarity_df = pd.DataFrame(columns=['search_id', 'match_id', 'similarity'])

for row_id in tqdm.tqdm(range(podcast_data.shape[0])):
    search_id = int(podcast_data.iloc[row_id]['id'])
    search_vec = lsi[corpus[row_id]]
    sims = index[search_vec]
    sims = sorted(enumerate(sims), key=lambda item: -item[1])
    
    # remove self match (should be first)
    for i, (id, sim) in enumerate(sims):
        if id == row_id:
            del sims[i]
            break
    
    # get match ids and similarities
    match_id = [podcast_data.iloc[match[0]]['id'] for match in sims]
    similarity = [match[1] for match in sims]
    
    similarity_df = similarity_df.append(pd.DataFrame({'search_id' : [[search_id] * len(match_id)],
                                                      'match_id' : [match_id],
                                                      'similarity' : [similarity]}))

similarity_df.to_pickle('gensim/similarity_df.pkl')

print index
