==> 09_Upload_Similarity_to_Database.ipynb <==
import psycopg2
import pandas as pd
import numpy as np

dbname = 'podcast'
username = 'lindsay'

similarity_df = pd.read_pickle('gensim/similarity_df.pkl')

con = psycopg2.connect(database = dbname, user = username)
cursor = con.cursor()

# unstack each list
match_id = similarity_df.match_id.apply(lambda x: pd.Series(x)).unstack()
search_id = similarity_df.search_id.apply(lambda x: pd.Series(x)).unstack()
similarity = similarity_df.similarity.apply(lambda x: pd.Series(x)).unstack()

# put unstacked lists in new df
similarity_unstacked = pd.DataFrame({'search_id' : search_id,
                                     'match_id' : match_id,
                                     'similarity' : similarity})

similarity_unstacked['match_id'] = [int(x) for x in similarity_unstacked['match_id']]

similarity_unstacked['similarity'] = [x.item() for x in similarity_unstacked['similarity']]

similarity_unstacked.dtypes

# insert similarity data 
query = "INSERT INTO podcast_similarity (search_id, match_id, similarity) VALUES (%s, %s, %s);"

for ind, row in similarity_unstacked.iterrows():
    data = (row['search_id'], row['match_id'], row['similarity'])
    cursor.execute(query, data)
    
con.commit()
